upd: fix UI 80%
