+++ src/App.js
@@ -5,7 +5,7 @@ import Home from "./components/Home";
 import Podcast from "./components/Podcast";
 import Blog from "./components/Blog";
 import FabHoiDap from "./components/FabHoiDap";
-import ChiTietBaiViet from "./components/Blog/ChiTietBaiViet";
+import ChiTietBaiViet from "./components/Blog/Components/ChiTietBaiViet";
 
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import { MainProvider } from "./contexts/MainContext";
